change : `TokenEstimator` moved to `token_estimators.py`.
